requirements.txt
fastapi==0.74.1
numpy==1.22.1
opencv_python==4.5.5.62
Pillow==9.0.1
tensorflow-hub==0.7.0
scikit_learn==1.0.2
starlette==0.17.1
tensorflow==2.8.0
torchvision==0.12.0
uvicorn==0.17.5
